Use new expenses data file format
The expenses data file format changed in 2017, but we were still using
the initial expenses file format, which needed to be amended before
loading the data.

With this changes we're now using the new expenses data file format,
which will save us additional work upon receiving new data files in the
future.